+++ src/app/api/_data/_mock/index.ts
@@ -2,4 +2,3 @@ export * from './_mock'
 export * from './assets'
 export * from './_others'
 export * from './_wallet'
-export * from './_countries'
