+++ src/components/country-select/country-select.tsx
@@ -3,7 +3,7 @@ import TextField from '@mui/material/TextField'
 import InputAdornment from '@mui/material/InputAdornment'
 import Autocomplete, { AutocompleteProps } from '@mui/material/Autocomplete'
 
-import { allCountries } from 'src/app/api/_data/_mock'
+import { allCountries } from 'src/app/api/_data/_countries'
 
 import Iconify from 'src/components/iconify'
