+++ src/sections/auth/jwt/jwt-login-view.tsx
@@ -27,7 +27,7 @@ import { getRecaptchaLng } from 'src/utils/format-number'
 
 import { useAuthContext } from 'src/auth/hooks'
 import { useLocales, useTranslate } from 'src/locales'
-import { allCountries } from 'src/app/api/_data/_mock'
+import { allCountries } from 'src/app/api/_data/_countries'
 import { PATH_AFTER_LOGIN, RECAPTCHA_SITE_KEY } from 'src/config-global'
 
 import FormProvider, { RHFCode, RHFTextField } from 'src/components/hook-form'
